requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
pydantic==2.5.0
python-docx==1.1.0
reportlab==4.0.7